Tasks 4.15-4.23 done

+++ Osa4/Bloglist/controllers/blogs.js
@@ -1,18 +1,24 @@
 const blogsRouter = require("express").Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const { tokenExtractor, userExtractor } = require('../utils/middleware')
+
 
 blogsRouter.get('/', async (request, response, next) => {
   try {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user', '-blogs')
     response.json(blogs)
   } catch (error) {
     next(error)
   }
 })
 
-blogsRouter.post('/', async (request, response, next) => {
+blogsRouter.post('/', tokenExtractor, userExtractor, async (request, response, next) => {
   try {
     const { title, author, url, likes } = request.body
+
+    const user = request.user
 
     if (title === undefined || url === undefined) {
       return response.status(400).json({error: 'title or url missing'})
@@ -22,25 +28,37 @@ blogsRouter.post('/', async (request, response, next) => {
       title,
       author,
       url,
-      likes: likes === undefined ? 0 : likes
+      likes: likes === undefined ? 0 : likes,
+      user: user._id
     })
 
     const savedBlog = await blog.save()
+    
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+
     response.json(savedBlog)
+
   } catch (error) {
     next(error)
   }
 })
 
-blogsRouter.delete('/:id', async (request, response, next) => {
+blogsRouter.delete('/:id', tokenExtractor, userExtractor, async (request, response, next) => {
   try {
-    const deletedBlog = await Blog.findByIdAndDelete(request.params.id)
+    const blogToDelete = await Blog.findById(request.params.id)
 
-    if (deletedBlog) {
-      response.status(204).end()
-    } else {
+    if (!blogToDelete) {
       response.status(404).json({ error: 'blog not found' })
     }
+
+    if (blogToDelete.user.toString() !== request.user._id.toString()) {
+      return response.status(403).json({ error: 'forbidden: not the blog owner' })
+    }
+
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+
   } catch (error) {
     next(error)
   }
